Update run_benchmark and create_report, add get bottleneck by perf and bench-runner

+++ run_benchmark.sh
@@ -4,40 +4,97 @@
 
 QCRAFT_TOP_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")/.." && pwd -P)"
 QBUILD_TOP_DIR="$QCRAFT_TOP_DIR/qbuild"
+
 # shellcheck disable=SC1090,SC1091
-. "${QCRAFT_TOP_DIR}"/scripts/shflags
-. "${QCRAFT_TOP_DIR}"/scripts/util/util.sh
+. "${QCRAFT_TOP_DIR}/scripts/shflags"
+
 # shellcheck disable=SC1090,SC1091
-source "${QCRAFT_TOP_DIR}/scripts/qcraft_base.sh"
+. "${QCRAFT_TOP_DIR}/scripts/qcraft_base.sh"
+
+# DEFINE_boolean 'local' "${FLAGS_TRUE}" 'local changes benchmark'
+DEFINE_boolean 'ci' "${FLAGS_FALSE}" 'benchmarking previous changes on ci'
+DEFINE_boolean 'cpuonly' "${FLAGS_FALSE}" 'bazel build on cpu only mode'
+DEFINE_string 'target_platform' "" 'Build a/o run at which platform, eg: x86_cpuonly, j5, x9hp, drive_orin, etc.'
+
 # shellcheck disable=SC1090,SC1091
 source "${QBUILD_TOP_DIR}/scripts/qbuild_utils.sh"
 
-# DEFINE_boolean 'local' "${FLAGS_FALSE}" 'local changes benchmark'
-# DEFINE_boolean 'ci' "${FLAGS_TRUE}" 'benchmarking previous changes on ci'
-
-# # get shflags from commandline(must called here)
-# FLAGS "$@" || exit $?
-# eval set -- "${FLAGS_ARGV}"
-
-# if [ "${FLAGS_local}" -eq ${FLAGS_TRUE} ]; then
-printf "On branch ${GREEN}%s${NO_COLOR} ${BLUE}(%s)${NO_COLOR}\n" "$(git rev-parse --abbrev-ref HEAD)" "$(git rev-parse --short HEAD)"
-info "qbuild: changes not staged for commit:"
-git diff --ignore-submodules --diff-filter=d --name-only # | grep -E '_bm|_test'
-#
-
-# if [ "${FLAGS_ci}" -eq ${FLAGS_TRUE} ]; then
-#   echo "benchmarking previous changes on ci"
-#   # clean
-# fi
-
-runs=2 # ci run benchmark per time
+runs=2
+PLATFORM="${FLAGS_target_platform}"
+platforms=(j5 x86_64_cpu x86_64_gpu x9hp drive_orin)
 current_commit=$(git rev-parse --short HEAD)
+BASE_COMMIT=
 RUN_OUTPUT_PATH="/tmp/bench_results"
 RUN_OUTPUT_FILE=
+BAZEL_FLAGS=
 
-if [[ ! -d $RUN_OUTPUT_PATH ]]; then
-  mkdir $RUN_OUTPUT_PATH
+bazel_flags=()
+
+if [ "${FLAGS_ci}" -eq ${FLAGS_TRUE} ]; then
+  if [[ $CI_MERGE_REQUEST_IID ]]; then
+    MR_ID=$CI_MERGE_REQUEST_IID
+  else
+    MR_ID=
+  fi
+else
+  MR_ID=
 fi
+
+function determine_build_flags() {
+  if [ -n "${BAZEL_FLAGS}" ]; then
+    bazel_flags+=("$BAZEL_FLAGS")
+  else
+    info "qbuild: ci has no additional bazel flags."
+  fi
+
+  if [[ "${PLATFORM}" == "x86_64_cpu" ]]; then
+    bazel_flags+=("--config=cpu")
+  elif [[ "${PLATFORM}" == "x86_64_gpu" ]]; then
+    bazel_flags+=("--config=gpu")
+  elif [[ "${PLATFORM}" == "j5" ]]; then
+    bazel_flags+=("--config=j5")
+  elif [[ "${PLATFORM}" == "x9hp" ]]; then
+    bazel_flags+=("--config=x9hp")
+  elif [[ "${PLATFORM}" == "drive_orin" ]]; then
+    bazel_flags+=("--config=drive_orin")
+  else
+    bazel_flags+=("--config=gpu")
+  fi
+
+  if [ "${FLAGS_cpuonly}" -eq ${FLAGS_TRUE} ]; then
+    PLATFORM="cpu"
+  fi
+
+  if [ ${#bazel_flags[@]} ]; then
+    info "qbuild: Bazel flags '${bazel_flags[*]}'"
+  else
+    info "qbuild: No bazel flags."
+  fi
+}
+
+function init() {
+  if [[ ! "${platforms[*]}" =~ $PLATFORM ]]; then
+    error "qbuild: $PLATFORM is not exists."
+    exit 1
+  fi
+
+  determine_build_flags
+
+  if [ "${FLAGS_ci}" -eq ${FLAGS_TRUE} ]; then
+    info "Regression benchmarking on ci ..."
+    printf "On branch ${GREEN}%s${NO_COLOR} ${BLUE}(%s)${NO_COLOR}\n" "$(git rev-parse --abbrev-ref HEAD)" "$(git rev-parse --short HEAD)"
+    info "qbuild: changes for commit($current_commit):"
+    git show --raw "$current_commit" | grep -E 'M|A' | awk -F 'M' '{print $2}' | awk '$1=$1' | uniq -u
+  else
+    printf "On branch ${GREEN}%s${NO_COLOR} ${BLUE}(%s)${NO_COLOR}\n" "$(git rev-parse --abbrev-ref HEAD)" "$(git rev-parse --short HEAD)"
+    info "qbuild: changes not staged for commit:"
+    git diff --ignore-submodules --diff-filter=d --name-only # | grep -E '_bm|_test'
+  fi
+
+  if [[ ! -d $RUN_OUTPUT_PATH ]]; then
+    mkdir $RUN_OUTPUT_PATH
+  fi
+}
 
 function _single_run() {
   build_target_name=$1
@@ -51,7 +108,7 @@ function _single_run() {
 
   if [[ $build_target_name == *bm ]]; then
     info "${GREEN}Pre-runing ""$build_target_name"" ...${NO_COLOR}"
-    bazel run -c opt "$build_target_name"
+    bazel run -c opt "${bazel_flags[*]}" --verbose_failures --spawn_strategy=local "$build_target_name"
 
     # for i in {1..2..1}; do # dont modified this num
     # info "${GREEN}Starting benchmark run $i/$runs:${NO_COLOR}"
@@ -59,15 +116,17 @@ function _single_run() {
       echo "" > "$RUN_OUTPUT_FILE"
     fi
 
-    bazel run -c opt "$build_target_name" 2>&1 | tee -a "$RUN_OUTPUT_FILE"
+    start=$(date +%s)
+    bazel run -c opt "${bazel_flags[*]}" --verbose_failures --spawn_strategy=local "$build_target_name" 2>&1 | tee -a "$RUN_OUTPUT_FILE"
+    end=$(date +%s)
     echo -e "\n\n" | tee -a "$RUN_OUTPUT_FILE"
-    info "RUN_OUTPUT_FILE: $RUN_OUTPUT_FILE"
+    info "RUN_OUTPUT_FILE: $RUN_OUTPUT_FILE, \n      Elapsed time: $((end - start))s"
     # done
 
     # MAKE SUMMARY
-    # python "${QBUILD_TOP_DIR}"/bench-rt/report/create_rt-report.py "$RUN_OUTPUT_FILE"
+    # python "${QBUILD_TOP_DIR}"/qbuild-bench/report/create_report.py "$RUN_OUTPUT_FILE"
   else
-    info "qbuild benchmark: Only ${BOLD}build${NO_COLOR} test, as '$build_target_name' is not gbench target."
+    info "qbuild benchmark: Only ${BOLD}build${NO_COLOR} test, as '$build_target_name' is not a gbench target."
     # bazel build -c opt --verbose_failures "$build_target_name"
   fi
 }
@@ -85,19 +144,19 @@ function _single_run_compare() {
 
   if [[ $build_target_name == *bm ]]; then
     info "${GREEN}Pre-runing ""$build_target_name"" ...${NO_COLOR}"
-    bazel run -c opt "$build_target_name"
+    bazel run -c opt "${bazel_flags[*]}" --verbose_failures --spawn_strategy=local "$build_target_name"
 
     for i in {1..2..1}; do # dont modified this num
       info "${GREEN}Starting benchmark run $i/$runs:${NO_COLOR}"
-      bazel run -c opt "$build_target_name" 2>&1 | tee -a "$RUN_OUTPUT_FILE"
+      bazel run -c opt "${bazel_flags[*]}" --verbose_failures --spawn_strategy=local "$build_target_name" 2>&1 | tee -a "$RUN_OUTPUT_FILE"
       echo -e "\n\n" | tee -a "$RUN_OUTPUT_FILE"
       info "RUN_OUTPUT_FILE: $RUN_OUTPUT_FILE"
     done
 
     # MAKE SUMMARY
-    python "${QBUILD_TOP_DIR}"/bench-rt/report/create_rt-report.py "$RUN_OUTPUT_FILE"
+    python "${QBUILD_TOP_DIR}"/qbuild-bench/report/create_report.py "$RUN_OUTPUT_FILE"
   else
-    info "qbuild benchmark: Only ${BOLD}build${NO_COLOR} test, as '$build_target_name' is not gbench target."
+    info "qbuild benchmark: Only ${BOLD}build${NO_COLOR} test, as '$build_target_name' is not a gbench target."
     # bazel build -c opt --verbose_failures "$build_target_name"
   fi
 }
@@ -128,11 +187,17 @@ function benchmark_regression() {
 
   rm -f $RUN_OUTPUT_PATH/*
 
-  # shellcheck disable=SC2178
-  # changed_files=$(_get_changed_files_local_diff)
-  changed_files=$(_get_changed_files_by_commit)
+  if [ "${FLAGS_ci}" -eq ${FLAGS_TRUE} ]; then
+    # shellcheck disable=SC2178
+    changed_files=$(_get_changed_files_by_commit)
+  else
+    # shellcheck disable=SC2178
+    changed_files=$(_get_changed_files_local_diff)
+  fi
+
   bazel_target_name=
 
+  no_affect_benchmark_flag='False'
   # shellcheck disable=SC2048
   for one_change in ${changed_files[*]}; do
     # echo "one_change: $one_change"
@@ -140,36 +205,69 @@ function benchmark_regression() {
     [[ $one_change == *.h ]] && bazel_target_name=${one_change%.h*}${one_change##*.h}    # || warning "qbuild: only support C"
 
     if [[ -n $bazel_target_name ]]; then
+      no_affect_benchmark_flag='True'
+
       # info "bazel build module_name: $bazel_target_name"
       dir_name=$(dirname "$bazel_target_name")
       base_file_name=$(basename "$bazel_target_name")
 
       AffectedBuildTarget=$(find_depend_target_by_file "$bazel_target_name")
 
-      info "${GREEN}Starting benchmark run ${NO_COLOR} ${BLUE}$current_commit${NO_COLOR}"
+      info "${GREEN}Starting benchmark run on ${NO_COLOR} ${BLUE}$current_commit${NO_COLOR}"
+
+      no_benchmark_test_flag='False'
       # shellcheck disable=SC2068
       for build_target in ${AffectedBuildTarget[@]}; do
-        info "Single run: $build_target ${GREEN}($current_commit)${NO_COLOR}"
-        _single_run "$build_target"
+        if [[ $build_target == *bm ]]; then
+          info "Single run: $build_target ${GREEN}($current_commit)${NO_COLOR}"
+          _single_run "$build_target"
+          no_benchmark_test_flag='True'
+        fi
       done
 
-      # checkout previous version
-      git reset --hard HEAD^
+      if [[ $no_benchmark_test_flag == 'False' ]]; then
+        info "Didn't affect any benchmark."
+        # python "${QCRAFT_TOP_DIR}"/qbuild/qbuild-bench/report/create_report.py --no_affect=1
+        break
+      fi
+
+      if [ "${FLAGS_ci}" -eq ${FLAGS_TRUE} ]; then
+        # checkout previous version
+        git reset --hard HEAD^
+        sleep 2s
+      else
+        git stash
+        # git checkout .
+      fi
       current_commit_tmp=$(git rev-parse --short HEAD)
-      info "${GREEN}Starting benchmark run ${NO_COLOR} ${BLUE}$current_commit_tmp${NO_COLOR}"
-      sleep 2s
+      BASE_COMMIT=$current_commit_tmp
+      info "${GREEN}Starting benchmark run on ${NO_COLOR} ${BLUE}$current_commit_tmp${NO_COLOR}"
+
       # shellcheck disable=SC2068
       for build_target in ${AffectedBuildTarget[@]}; do
+        if ! bazel query "$build_target"; then
+          warning "qbuild: base version($current_commit_tmp) has no $build_target."
+          continue
+        fi
         info "Single run: $build_target ${GREEN}($current_commit_tmp)${NO_COLOR}"
         _single_run "$build_target"
       done
 
-      git reset --hard "$current_commit"
+      if [ "${FLAGS_ci}" -eq ${FLAGS_TRUE} ]; then
+        git reset --hard "$current_commit"
+      else
+        git stash pop
+      fi
     fi
   done
 
+  if [[ $no_affect_benchmark_flag == 'False' ]]; then
+    info "Your mr didn't affect any benchmark."
+  fi
+
   # MAKE SUMMARY
-  python "${QBUILD_TOP_DIR}"/bench-rt/report/create_rt-report.py "$RUN_OUTPUT_PATH/"
+  info "$RUN_OUTPUT_PATH/ --mr=$MR_ID --commit_sha=$current_commit --base_commit=$BASE_COMMIT"
+  python "${QBUILD_TOP_DIR}"/qbuild-bench/report/create_report.py --file_path="$RUN_OUTPUT_PATH/" --mr="$MR_ID" --commit_sha="$current_commit" --base_commit="$BASE_COMMIT"
 }
 
 function run_increase_file2target() {
@@ -201,22 +299,22 @@ function run_increase_file2target() {
           qbuild --run j5 "$bazel_target_name"
         else # run on x86
           if [[ $base_file_name == *test ]]; then
-            bazel run "//$dir_name:$base_file_name"
+            bazel test -c opt --verbose_failures --spawn_strategy=local "//$dir_name:$base_file_name"
           fi
           if [[ $base_file_name == *bm ]]; then
             # shellcheck disable=SC2140
             info "${GREEN}Pre-runing "//"$dir_name":"$base_file_name"" ...${NO_COLOR}"
-            bazel run "//$dir_name:$base_file_name"
+            bazel run -c opt "${bazel_flags[*]}" --verbose_failures --spawn_strategy=local "//$dir_name:$base_file_name"
 
             for i in {1..2..1}; do # dont modified this num
               info "${GREEN}Starting benchmark run $i/$runs:${NO_COLOR}"
-              bazel run "//$dir_name:$base_file_name" 2>&1 | tee -a "$RUN_OUTPUT_FILE"
+              bazel run -c opt "${bazel_flags[*]}" --verbose_failures --spawn_strategy=local "//$dir_name:$base_file_name" 2>&1 | tee -a "$RUN_OUTPUT_FILE"
               echo -e "\n\n" | tee -a "$RUN_OUTPUT_FILE"
               info "RUN_OUTPUT_FILE: $RUN_OUTPUT_FILE"
             done
 
             # MAKE SUMMARY
-            python "${QBUILD_TOP_DIR}"/bench-rt/report/create_rt-report.py "$RUN_OUTPUT_FILE"
+            python "${QBUILD_TOP_DIR}"/qbuild-bench/report/create_report.py "$RUN_OUTPUT_FILE"
           fi
         fi
       else
@@ -234,15 +332,23 @@ function run_increase_file2target() {
 }
 
 function main() {
-  # local platform=${1:-"x86"}
-  # run_increase_file2target "$platform"
+  # get shflags from commandline(must called here)
+  FLAGS "$@" || exit $?
+  eval set -- "${FLAGS_ARGV}"
+
+  init
 
   benchmark_regression
+
+  # local platform=${1:-"x86"}
+  # run_increase_file2target "$platform"
 
   # _single_run "//onboard/math:fast_math_bm"
 
   # find_depend_target_by_file onboard/lite/lite_timer.h
   # _get_changed_files_by_commit
+
+  info "Run Successfully!"
 }
 
 main "$@"
